+++ flutter_inappwebview_platform_interface/lib/src/platform_webview_feature.g.dart
@@ -23,6 +23,12 @@ class WebViewFeature {
     'ALGORITHMIC_DARKENING',
   );
 
+  ///Feature for [isFeatureSupported]. This feature covers [InAppWebViewSettings.attributionRegistrationBehavior].
+  static const ATTRIBUTION_REGISTRATION_BEHAVIOR = WebViewFeature._internal(
+    'ATTRIBUTION_REGISTRATION_BEHAVIOR',
+    'ATTRIBUTION_REGISTRATION_BEHAVIOR',
+  );
+
   ///Feature for [isFeatureSupported]. This feature covers [InAppWebViewController.createWebMessageChannel].
   static const CREATE_WEB_MESSAGE_CHANNEL = WebViewFeature._internal(
     'CREATE_WEB_MESSAGE_CHANNEL',
@@ -332,6 +338,7 @@ class WebViewFeature {
   ///Set of all values of [WebViewFeature].
   static final Set<WebViewFeature> values = [
     WebViewFeature.ALGORITHMIC_DARKENING,
+    WebViewFeature.ATTRIBUTION_REGISTRATION_BEHAVIOR,
     WebViewFeature.CREATE_WEB_MESSAGE_CHANNEL,
     WebViewFeature.DISABLED_ACTION_MODE_MENU_ITEMS,
     WebViewFeature.DOCUMENT_START_SCRIPT,
@@ -463,6 +470,8 @@ class WebViewFeature {
     switch (_value) {
       case 'ALGORITHMIC_DARKENING':
         return 'ALGORITHMIC_DARKENING';
+      case 'ATTRIBUTION_REGISTRATION_BEHAVIOR':
+        return 'ATTRIBUTION_REGISTRATION_BEHAVIOR';
       case 'CREATE_WEB_MESSAGE_CHANNEL':
         return 'CREATE_WEB_MESSAGE_CHANNEL';
       case 'DISABLED_ACTION_MODE_MENU_ITEMS':
@@ -602,6 +611,13 @@ class AndroidWebViewFeature {
     'ALGORITHMIC_DARKENING',
     'ALGORITHMIC_DARKENING',
   );
+
+  ///Feature for [isFeatureSupported]. This feature covers [InAppWebViewSettings.attributionRegistrationBehavior].
+  static const ATTRIBUTION_REGISTRATION_BEHAVIOR =
+      AndroidWebViewFeature._internal(
+        'ATTRIBUTION_REGISTRATION_BEHAVIOR',
+        'ATTRIBUTION_REGISTRATION_BEHAVIOR',
+      );
 
   ///
   static const CREATE_WEB_MESSAGE_CHANNEL = AndroidWebViewFeature._internal(
@@ -888,6 +904,7 @@ class AndroidWebViewFeature {
   ///Set of all values of [AndroidWebViewFeature].
   static final Set<AndroidWebViewFeature> values = [
     AndroidWebViewFeature.ALGORITHMIC_DARKENING,
+    AndroidWebViewFeature.ATTRIBUTION_REGISTRATION_BEHAVIOR,
     AndroidWebViewFeature.CREATE_WEB_MESSAGE_CHANNEL,
     AndroidWebViewFeature.DISABLED_ACTION_MODE_MENU_ITEMS,
     AndroidWebViewFeature.DOCUMENT_START_SCRIPT,
@@ -1015,6 +1032,8 @@ class AndroidWebViewFeature {
     switch (_value) {
       case 'ALGORITHMIC_DARKENING':
         return 'ALGORITHMIC_DARKENING';
+      case 'ATTRIBUTION_REGISTRATION_BEHAVIOR':
+        return 'ATTRIBUTION_REGISTRATION_BEHAVIOR';
       case 'CREATE_WEB_MESSAGE_CHANNEL':
         return 'CREATE_WEB_MESSAGE_CHANNEL';
       case 'DISABLED_ACTION_MODE_MENU_ITEMS':